job/requirements.txt
kaggle>=1.6.0
google-cloud-storage==2.18.2
